PP-10485 Rename IT classes
It's easier for discovery and IDE navigation if test class names contain
the name of the resource they're testing. We follow this pattern for
most of our integration test classes - update IT test classes that don't
follow this pattern.
