+++ src/test/java/uk/gov/pay/api/it/PaymentsResourceSearchIT.java
@@ -47,7 +47,7 @@ import static uk.gov.pay.api.it.fixtures.PaymentSearchResultBuilder.DEFAULT_SETT
 import static uk.gov.pay.api.it.fixtures.PaymentSearchResultBuilder.aSuccessfulSearchPayment;
 import static uk.gov.pay.api.utils.Urls.paymentLocationFor;
 
-public class PaymentResourceSearchIT extends PaymentResourceITestBase {
+public class PaymentsResourceSearchIT extends PaymentResourceITestBase {
 
     private static final String TEST_REFERENCE = "test_reference";
     private static final String TEST_EMAIL = "alice.111@mail.fake";
